Add separate file for functions. Modify 'index.php' for working with 'fce.php'
