+++ www/index.php
@@ -1,5 +1,6 @@
 <?php
 declare(strict_types=1);
+require 'fce.php';
 
 $items = [
     0 => "Lorem ipsum dolor sit amet, consectetuer adipiscing elit.",
@@ -13,9 +14,7 @@ $items = [
     8 => "Etiam posuere lacus quis dolor.",
 ];
 
-
 printNiceHtmlHeader();
-
 
 $id = getId();
 
@@ -23,46 +22,4 @@ if ($id !== null) {
     printValueFromItems($items, $id);
 } else {
     printForm();
-}
-
-
-function printNiceHtmlHeader(): void
-{
-    echo file_get_contents('./assets/nice-html.html');
-}
-
-
-function printValueFromItems(array $items, int $id): void
-{
-    try {
-        echo "Byla vybrána tato položka: " . getValueFromItems($items, $id);
-    } catch (Exception $e) {
-        echo 'Chyba: ' . $e->getMessage();
-    }
-}
-
-
-function printForm(): void
-{
-    echo '<form method="get">Zadejte číslo: <input name="id"><input type="submit">';
-}
-
-
-function getValueFromItems(array $items, int $id): string
-{
-    if (isset($items[$id])) {
-        return $items[$id];
-    }
-
-    throw new Exception('Zadali jste neexistující položku');
-}
-
-
-function getId(): ?int
-{
-    if (isset($_GET['id']) && is_numeric($_GET['id'])) {
-        return (int)$_GET['id'];
-    }
-
-    return null;
 }
